feat(mgb69-alter-new-appointment): create new types and organize code for appointments
