+++ @types/barberScheduling.ts
@@ -1,0 +1,7 @@
+export interface BarberSchedulingData {
+  id: string;
+  barberId: string;
+  dayAt: string;
+  startTime: string;
+  endTime: string
+}
