+++ @types/form.ts
@@ -1,0 +1,6 @@
+export interface FormData {
+  barberId: string;
+  dayAt: Date;
+  hourAt: string;
+  serviceType: string;
+}
