+++ @types/scheduling.ts
@@ -1,3 +1,13 @@
+export interface Scheduling {
+  id: string;
+  userId: string;
+  barberId: string;
+  dayAt: string;
+  hourAt: string;
+  serviceType: string;
+  status: string;
+}
+
 export interface SchedulingData {
   id: string;
   userId: string;
